pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["metamaterial_rl_env"]

[project]
name = "metamaterial_rl_env"
version = "0.0.1"
dependencies = [
  "pygame>=2.1.3",
  "torchrl",
  "numpy",
  "numba",
  "matplotlib",
  "pre-commit",
]
